+++ pipeline/snapshot.go
@@ -15,7 +15,7 @@ func (p *Pipeline) sendSnapshots(ctx context.Context, storeMap store.Map) error 
 	if reqctx.Details(ctx).IsSubRequest {
 		return nil
 	}
-	snapshotModules := reqctx.Details(ctx).Request.InitialStoreSnapshotForModules
+	snapshotModules := reqctx.Details(ctx).Request.DebugInitialStoreSnapshotForModules
 	if len(snapshotModules) == 0 {
 		return nil
 	}